requirements.txt
tensorflow
numpy
pandas
matplotlib
scikit-learn
pillow
opencv-python